move validation tests to NodeViewCustomization so that CoreNodeWPF is loaded at test time.
