+++ test/DynamoCoreWpfTests/NodeViewCustomizationTests.cs
@@ -15,6 +15,8 @@ using Dynamo.Utilities;
 using DynamoCoreWpfTests.Utility;
 using NUnit.Framework;
 using CoreNodeModels.Input;
+using CoreNodeModelsWpf.Nodes;
+using System.Globalization;
 
 namespace DynamoCoreWpfTests
 {
@@ -134,6 +136,41 @@ namespace DynamoCoreWpfTests
             Assert.IsTrue(number.IsInErrorState);
 
         }
+
+        //these tests live here as they have a dependencey on CoreNodesWpf which
+        //is only accessible at test time after it is loaded via Dynamo 
+        [Test]
+        [Category("UnitTests")]
+        public static void validateNumericFailsOnNonNumericInputs()
+        {
+            var numericalRule = new NumericValidationRule();
+
+            Assert.IsFalse(numericalRule.Validate("0..10", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("0..10..2", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("0..10..5", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("0..10..#5", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("0.0..10..0.5", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("start..end", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("a..b", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("a..10", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsFalse(numericalRule.Validate("a", CultureInfo.InvariantCulture).IsValid);
+
+        }
+        [Test]
+        [Category("UnitTests")]
+        public static void validateNumericPassesOnNumericInputs()
+        {
+            var numericalRule = new NumericValidationRule();
+
+            Assert.IsTrue(numericalRule.Validate("10", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsTrue(numericalRule.Validate("2147483647", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsTrue(numericalRule.Validate("9223372036854775807", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsTrue(numericalRule.Validate(".00000000001", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsTrue(numericalRule.Validate("100,000,000", CultureInfo.InvariantCulture).IsValid);
+            Assert.IsTrue(numericalRule.Validate("1,2", CultureInfo.InvariantCulture).IsValid);
+
+        }
+
 
         [Test]
         public void IntegerSliderHasSliderAndCorrectValues()
